Avoid marking coordinate as unknown when parsing a M18/M84 S<timeout>
M18/M84 commands are used to disable steppers, however they also
are used to set the stepper inactivity timeout:

    M18 S60; Set the stepper inactivity timeout to 1 minute

When only setting the timeout, steppers are not disabled, so there
is not need to mark the coordinates as unknown.

To fix this, I added a simple test, in the case a M18/M84 command
is seen, it checks for the arguments and if only S is specified,
it will not mark coordinates as unknown.